feat: Add QNAP runtime Dockerfile and build scripts for ThemisDB
- Introduced Dockerfile.qnap.runtime for building a QNAP-compatible runtime image.
- Added build-qnap.sh script for building ThemisDB on QNAP with vcpkg integration.
- Created RELEASE_DISTRIBUTION_STRATEGY.md to outline the release and distribution process.
- Implemented version bumping script (bump_version.ps1) for managing semantic versioning.
- Developed packaging script (package_releases.ps1) for creating release artifacts.
- Initialized VERSION file to track the current version of ThemisDB.

+++ src/query/semantic_cache.cpp
@@ -424,8 +424,9 @@ SemanticQueryCache::Status SemanticQueryCache::saveCacheEntry_(const CacheEntry&
         entry.created_at.time_since_epoch()).count());
     entity.setField("last_accessed", std::chrono::duration_cast<std::chrono::milliseconds>(
         entry.last_accessed.time_since_epoch()).count());
-    entity.setField("hit_count", entry.hit_count);
-    entity.setField("result_size", static_cast<int>(entry.result_size));
+    // Ensure integral fields match Value's int64_t alternative
+    entity.setField("hit_count", static_cast<int64_t>(entry.hit_count));
+    entity.setField("result_size", static_cast<int64_t>(entry.result_size));
     
     db_.put(key, entity.serialize());
